requirements.txt
alembic==1.13.3
aniso8601==9.0.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
elastic-transport==8.15.1
elasticsearch==7.13.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.5
MarkupSafe==3.0.1
mysqlclient==2.2.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
zipp==3.20.2
